requirements.txt
streamlit==1.31.1
chromadb==0.4.22
sentence-transformers==2.6.1
pymupdf==1.23.14
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.4
